+++ src/server/src/projects/services/projects.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { TypeOrmCrudService } from "@nestjsx/crud-typeorm";
-import { Repository } from "typeorm";
+import { Repository, SelectQueryBuilder } from "typeorm";
 
 import { Project } from "../entities/project.entity";
 import { ProjectVisibility } from "../../../../shared/constants";
@@ -10,6 +10,7 @@ import { paginate, Pagination, IPaginationOptions } from "nestjs-typeorm-paginat
 type AllProjectsOptions = IPaginationOptions & {
   readonly completed?: boolean;
   readonly region?: string;
+  readonly userId?: string;
 };
 
 @Injectable()
@@ -22,16 +23,12 @@ export class ProjectsService extends TypeOrmCrudService<Project> {
     return this.repo.save(project);
   }
 
-  async findAllPublishedProjectsPaginated(
-    options: AllProjectsOptions
-  ): Promise<Pagination<Project>> {
-    // Returns admin-only listing of all organization projects
-    const builder = this.repo
+  getProjectsBase(): SelectQueryBuilder<Project> {
+    return this.repo
       .createQueryBuilder("project")
       .innerJoinAndSelect("project.regionConfig", "regionConfig")
       .innerJoinAndSelect("project.user", "user")
       .leftJoinAndSelect("project.chamber", "chamber")
-      .where("project.visibility = :published", { published: ProjectVisibility.Published })
       .select([
         "project.id",
         "project.numberOfDistricts",
@@ -44,6 +41,14 @@ export class ProjectsService extends TypeOrmCrudService<Project> {
         "user.name"
       ])
       .orderBy("project.updatedDt", "DESC");
+  }
+
+  async findAllPublishedProjectsPaginated(
+    options: AllProjectsOptions
+  ): Promise<Pagination<Project>> {
+    const builder = this.getProjectsBase().andWhere("project.visibility = :published", {
+      published: ProjectVisibility.Published
+    });
     const builderWithFilter = options.completed
       ? // Completed projects are defined as having no population in the unassigned district
         builder.andWhere(
@@ -54,5 +59,17 @@ export class ProjectsService extends TypeOrmCrudService<Project> {
       ? builderWithFilter.andWhere("regionConfig.regionCode = :region", { region: options.region })
       : builderWithFilter;
     return paginate<Project>(builderWithRegion, options);
+  }
+
+  async findAllUserProjectsPaginated(
+    userId: string,
+    options: AllProjectsOptions
+  ): Promise<Pagination<Project>> {
+    const builder = this.getProjectsBase().andWhere(
+      "project.archived = FALSE AND user.id = :userId",
+      { userId }
+    );
+
+    return paginate<Project>(builder, options);
   }
 }